feat(access): PRET-6 stage A — the flip subsystem retires; the readiness scan re-homes
access-enforcement.ts + auto-flip-pass.ts + pret-flip-estimate.mjs deleted; the blind-principal
+ unpartitioned scans survive as lib/admin/access-health.ts (assessAccessHealth — same oracle-
derived checks, lockout framing) with the CLI access-health command; the scheduler's flip slot
deleted (the pret4 materialization retry slot SURVIVES, per the cold read); the inspector sheds
its autoFlip stuck-state field; docker bootstrap drains directly via the new drain-context CLI
(the PRET-2 cold-read-M2 fix kept without the flip machinery); spec-first precondition suite
4/4 green
